pyproject.toml
[project]
name = "dadd-e"
version = "0.1.0"
description = "Dadd-E: A real-time productivity assistant using Omi glasses"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Omi Integration
    "omi-sdk>=0.2.1",
    "aiohttp>=3.11.14",
    # Voice & Audio
    "deepgram-sdk==2.12.0",
    "websockets>=13.1",
    # AI & Vision
    "openai>=1.54.0",
    "anthropic>=0.39.0",
    # Database & Storage
    "supabase>=2.10.0",
    "redis>=5.2.0",
    "asyncpg>=0.30.0",
    # Utilities
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pillow>=11.0.0",
    "numpy>=2.1.0",
    "composio-core==0.6.10",
    "composio-openai==0.6.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio",
    "black>=24.10.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
